project/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.5.1
anyio=4.7.0=py310hca03da5_0
appnope=0.1.4=py310hca03da5_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h80987f9_1
asttokens=3.0.0=py310hca03da5_0
async-lru=2.0.4=py310hca03da5_0
attrs=24.3.0=py310hca03da5_0
babel=2.16.0=py310hca03da5_0
beautifulsoup4=4.13.4=py310hca03da5_0
bleach=6.2.0=py310hca03da5_0
brotlicffi=1.0.9.2=py310h313beb8_1
bzip2=1.0.8=h80987f9_6
ca-certificates=2025.7.15=hca03da5_0
certifi=2025.8.3=py310hca03da5_0
cffi=1.17.1=py310h3eb5a62_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
comm=0.2.1=py310hca03da5_0
debugpy=1.8.11=py310h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
exceptiongroup=1.2.0=py310hca03da5_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.7.1=h313beb8_0
h11=0.16.0=py310hca03da5_0
httpcore=1.0.9=py310hca03da5_0
httpx=0.28.1=py310hca03da5_0
idna=3.7=py310hca03da5_0
ipykernel=6.29.5=py310hca03da5_1
ipython=8.30.0=py310hca03da5_0
jedi=0.19.2=py310hca03da5_0
jinja2=3.1.6=py310hca03da5_0
json5=0.9.25=py310hca03da5_0
jsonschema=4.25.0=py310hca03da5_0
jsonschema-specifications=2023.7.1=py310hca03da5_0
jupyter-lsp=2.2.5=py310hca03da5_0
jupyter_client=8.6.3=py310hca03da5_0
jupyter_core=5.8.1=py310hca03da5_0
jupyter_events=0.12.0=py310hca03da5_0
jupyter_server=2.16.0=py310hca03da5_0
jupyter_server_terminals=0.5.3=py310hca03da5_0
jupyterlab=4.4.4=py310hca03da5_0
jupyterlab_pygments=0.3.0=py310hca03da5_0
jupyterlab_server=2.27.3=py310hca03da5_0
libcxx=19.1.7=hb09ecce_3
libffi=3.4.4=hca03da5_1
libsodium=1.0.18=h1a28f6b_0
markupsafe=3.0.2=py310h80987f9_0
matplotlib-inline=0.1.6=py310hca03da5_0
mistune=3.1.2=py310hca03da5_0
nbclient=0.10.2=py310hca03da5_0
nbconvert-core=7.16.6=py310hca03da5_0
nbformat=5.10.4=py310hca03da5_0
ncurses=6.5=hee39554_0
nest-asyncio=1.6.0=py310hca03da5_0
notebook-shim=0.2.4=py310hca03da5_0
numpy=2.2.6=pypi_0
openssl=3.0.17=h4ee41c1_0
overrides=7.4.0=py310hca03da5_0
packaging=25.0=py310hca03da5_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.4=py310hca03da5_0
pexpect=4.9.0=py310hca03da5_0
pip=25.1=pyhc872135_2
platformdirs=4.3.7=py310hca03da5_0
prometheus_client=0.21.1=py310hca03da5_0
prompt-toolkit=3.0.43=py310hca03da5_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=py310h80987f9_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.19.1=py310hca03da5_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.18=h19e8193_0
python-dateutil=2.9.0post0=py310hca03da5_2
python-dotenv=1.1.1=pypi_0
python-fastjsonschema=2.20.0=py310hca03da5_0
python-json-logger=3.2.1=py310hca03da5_0
pyyaml=6.0.2=py310h80987f9_0
pyzmq=26.2.0=py310h313beb8_0
readline=8.3=h0b18652_0
referencing=0.30.2=py310hca03da5_0
requests=2.32.4=py310hca03da5_0
rfc3339-validator=0.1.4=py310hca03da5_0
rfc3986-validator=0.1.1=py310hca03da5_0
rpds-py=0.22.3=py310h2aea54e_0
send2trash=1.8.2=py310hca03da5_1
setuptools=78.1.1=py310hca03da5_0
six=1.17.0=py310hca03da5_0
sniffio=1.3.0=py310hca03da5_0
soupsieve=2.5=py310hca03da5_0
sqlite=3.50.2=h79febb2_1
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.17.1=py310hca03da5_0
tinycss2=1.4.0=py310hca03da5_0
tk=8.6.15=hcd8a7d5_0
tomli=2.2.1=py310hca03da5_0
tornado=6.5.1=py310h80987f9_0
traitlets=5.14.3=py310hca03da5_0
typing-extensions=4.12.2=py310hca03da5_0
typing_extensions=4.12.2=py310hca03da5_0
tzdata=2025b=h04d1e81_0
urllib3=2.5.0=py310hca03da5_0
wcwidth=0.2.13=py310hca03da5_0
webencodings=0.5.1=py310hca03da5_1
websocket-client=1.8.0=py310hca03da5_0
wheel=0.45.1=py310hca03da5_0
xz=5.6.4=h80987f9_1
yaml=0.2.5=h1a28f6b_0
zeromq=4.3.5=h313beb8_0
zlib=1.2.13=h18a0788_1
